adding pieces for both players in game model
